+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added option to allow run post startup commands (`postStart`)
 - Added option to run the Kubernetes Pods without isolating the control network (i.e., network provided by Kubernets, usually `eth0`) keeping the control net in the root netns (`isolateControlNet`)
 - Fixed issue with bind9 mnsec service for the regex to validade zone being added
+- Added support for sysctls definition on Kubernetes Pods
 
 ## [1.1.0] - 2025-01-10